+++ pennylane_lightning/core/src/simulators/lightning_tensor/tncuda/TNCudaBase.hpp
@@ -20,12 +20,14 @@
 #pragma once
 
 #include <complex>
+#include <memory>
 #include <type_traits>
 #include <vector>
 
 #include <cuda.h>
 #include <cutensornet.h>
 
+#include "TNCudaGateCache.hpp"
 #include "TensorBase.hpp"
 #include "TensorCuda.hpp"
 #include "TensornetBase.hpp"
@@ -38,16 +40,24 @@ namespace {
 namespace cuUtil = Pennylane::LightningGPU::Util;
 using namespace Pennylane::LightningGPU;
 using namespace Pennylane::LightningTensor::TNCuda;
+using namespace Pennylane::LightningTensor::TNCuda::Gates;
 using namespace Pennylane::LightningTensor::TNCuda::Util;
 } // namespace
 ///@endcond
 
 namespace Pennylane::LightningTensor::TNCuda {
-
-template <class Precision, class Derived>
-class TNCudaBase : public TensornetBase<Precision, Derived> {
+/**
+ * @brief CRTP-enabled base class for cuTensorNet backends.
+ *
+ * @tparam PrecisionT Floating point precision.
+ * @tparam Derived Derived class to instantiate using CRTP.
+ */
+template <class PrecisionT, class Derived>
+class TNCudaBase : public TensornetBase<PrecisionT, Derived> {
   private:
-    using BaseType = TensornetBase<Precision, Derived>;
+    using CFP_t = decltype(cuUtil::getCudaType(PrecisionT{}));
+    using ComplexT = std::complex<PrecisionT>;
+    using BaseType = TensornetBase<PrecisionT, Derived>;
     SharedTNCudaHandle handle_;
     cudaDataType_t typeData_;
     DevTag<int> dev_tag_;
@@ -57,16 +67,19 @@ class TNCudaBase : public TensornetBase<Precision, Derived> {
         CUTENSORNET_STATE_PURITY_PURE; // Only supports pure tensor network
                                        // states as v24.03
 
+    std::shared_ptr<TNCudaGateCache<PrecisionT>> gate_cache_;
+
   public:
     TNCudaBase() = delete;
 
     explicit TNCudaBase(const std::size_t numQubits, int device_id = 0,
                         cudaStream_t stream_id = 0)
         : BaseType(numQubits), handle_(make_shared_tncuda_handle()),
-          dev_tag_({device_id, stream_id}) {
+          dev_tag_({device_id, stream_id}),
+          gate_cache_(std::make_shared<TNCudaGateCache<PrecisionT>>(dev_tag_)) {
         // TODO this code block could be moved to base class and need to revisit
         // when working on copy ctor
-        if constexpr (std::is_same_v<Precision, double>) {
+        if constexpr (std::is_same_v<PrecisionT, double>) {
             typeData_ = CUDA_C_64F;
             typeCompute_ = CUTENSORNET_COMPUTE_64F;
         } else {
@@ -87,10 +100,11 @@ class TNCudaBase : public TensornetBase<Precision, Derived> {
 
     explicit TNCudaBase(const std::size_t numQubits, DevTag<int> dev_tag)
         : BaseType(numQubits), handle_(make_shared_tncuda_handle()),
-          dev_tag_(dev_tag) {
+          dev_tag_(dev_tag),
+          gate_cache_(std::make_shared<TNCudaGateCache<PrecisionT>>(dev_tag_)) {
         // TODO this code block could be moved to base class and need to revisit
         // when working on copy ctor
-        if constexpr (std::is_same_v<Precision, double>) {
+        if constexpr (std::is_same_v<PrecisionT, double>) {
             typeData_ = CUDA_C_64F;
             typeCompute_ = CUTENSORNET_COMPUTE_64F;
         } else {
@@ -141,6 +155,128 @@ class TNCudaBase : public TensornetBase<Precision, Derived> {
         return dev_tag_;
     }
 
+    /**
+     * @brief Append multiple gates to the compute graph.
+     * NOTE: This function does not update the quantum state but only appends
+     * gate tensor operator to the graph.
+     * @param ops Vector of gate names to be applied in order.
+     * @param ops_wires Vector of wires on which to apply index-matched gate
+     * name.
+     * @param ops_adjoint Indicates whether gate at matched index is to be
+     * inverted.
+     * @param ops_params Vector of gate parameters.
+     */
+    void
+    applyOperations(const std::vector<std::string> &ops,
+                    const std::vector<std::vector<std::size_t>> &ops_wires,
+                    const std::vector<bool> &ops_adjoint,
+                    const std::vector<std::vector<PrecisionT>> &ops_params) {
+        const std::size_t numOperations = ops.size();
+        PL_ABORT_IF_NOT(
+            numOperations == ops_wires.size(),
+            "Invalid arguments: number of operations, wires, and inverses "
+            "must all be equal");
+        PL_ABORT_IF_NOT(
+            numOperations == ops_adjoint.size(),
+            "Invalid arguments: number of operations, wires and inverses"
+            "must all be equal");
+        for (std::size_t i = 0; i < numOperations; i++) {
+            this->applyOperation(ops[i], ops_wires[i], ops_adjoint[i],
+                                 ops_params[i]);
+        }
+    }
+
+    /**
+     * @brief Append multiple gate tensors to the compute graph.
+     * NOTE: This function does not update the quantum state but only appends
+     * gate tensor operator to the graph.
+     * @param ops Vector of gate names to be applied in order.
+     * @param ops_wires Vector of wires on which to apply index-matched gate
+     * name.
+     * @param ops_adjoint Indicates whether gate at matched index is to be
+     * inverted.
+     */
+    void applyOperations(const std::vector<std::string> &ops,
+                         const std::vector<std::vector<std::size_t>> &ops_wires,
+                         const std::vector<bool> &ops_adjoint) {
+        const std::size_t numOperations = ops.size();
+        PL_ABORT_IF_NOT(
+            numOperations == ops_wires.size(),
+            "Invalid arguments: number of operations, wires, and inverses "
+            "must all be equal");
+        PL_ABORT_IF_NOT(
+            numOperations == ops_adjoint.size(),
+            "Invalid arguments: number of operations, wires and inverses"
+            "must all be equal");
+        for (std::size_t i = 0; i < numOperations; i++) {
+            this->applyOperation(ops[i], ops_wires[i], ops_adjoint[i], {});
+        }
+    }
+
+    /**
+     * @brief Append a single gate tensor to the compute graph.
+     * NOTE: This function does not update the quantum state but only appends
+     * gate tensor operator to the graph.
+     * @param opName Gate's name.
+     * @param wires Wires to apply gate to.
+     * @param adjoint Indicates whether to use adjoint of gate.
+     * @param params Optional parameter list for parametric gates.
+     * @param gate_matrix Optional gate matrix for custom gates.
+     */
+    void applyOperation(const std::string &opName,
+                        const std::vector<std::size_t> &wires,
+                        bool adjoint = false,
+                        const std::vector<PrecisionT> &params = {0.0},
+                        const std::vector<ComplexT> &gate_matrix = {}) {
+        auto &&par = (params.empty()) ? std::vector<PrecisionT>{0.0} : params;
+        DataBuffer<PrecisionT, int> dummy_device_data(
+            Pennylane::Util::exp2(wires.size()), getDevTag());
+        int64_t id;
+        std::vector<int32_t> stateModes(wires.size());
+        std::transform(
+            wires.begin(), wires.end(), stateModes.begin(), [&](std::size_t x) {
+                return static_cast<int32_t>(BaseType::getNumQubits() - 1 - x);
+            });
+
+        // TODO: Need changes to support to the controlled gate tensor API once
+        // the API is finalized in cutensornet lib.
+        //  Note `adjoint` in the cutensornet context indicates whether or not
+        //  all tensor elements of the tensor operator will be complex
+        //  conjugated. `adjoint` in the following API is not equivalent to
+        //  `inverse` in the lightning context
+        PL_CUTENSORNET_IS_SUCCESS(cutensornetStateApplyTensorOperator(
+            /* const cutensornetHandle_t */ getTNCudaHandle(),
+            /* cutensornetState_t */ getQuantumState(),
+            /* int32_t numStateModes */ stateModes.size(),
+            /* const int32_t * stateModes */ stateModes.data(),
+            /* void * */ static_cast<void *>(dummy_device_data.getData()),
+            /* const int64_t *tensorModeStrides */ nullptr,
+            /* const int32_t immutable */ 1,
+            /* const int32_t adjoint */ adjoint,
+            /* const int32_t unitary */ 1,
+            /* int64_t * */ &id));
+        if (!gate_matrix.empty()) {
+            std::vector<CFP_t> matrix_cu(gate_matrix.size());
+            std::transform(gate_matrix.begin(), gate_matrix.end(),
+                           matrix_cu.begin(), [](const ComplexT &x) {
+                               return cuUtil::complexToCu<ComplexT>(x);
+                           });
+            auto gate_key = std::make_pair(opName, par);
+            gate_cache_->add_gate(static_cast<std::size_t>(id), gate_key,
+                                  matrix_cu);
+        } else {
+            gate_cache_->add_gate(static_cast<std::size_t>(id), opName, par);
+        }
+        PL_CUTENSORNET_IS_SUCCESS(cutensornetStateUpdateTensorOperator(
+            /* const cutensornetHandle_t */ getTNCudaHandle(),
+            /* cutensornetState_t */ getQuantumState(),
+            /* int64_t tensorId*/ id,
+            /* void* */
+            static_cast<void *>(
+                gate_cache_->get_gate_device_ptr(static_cast<std::size_t>(id))),
+            /* int32_t unitary*/ 1));
+    }
+
   protected:
     /**
      * @brief Returns the workspace size.
@@ -174,7 +310,7 @@ class TNCudaBase : public TensornetBase<Precision, Derived> {
      * @param worksize Memory size of a work space
      */
     void setWorkSpaceMemory(cutensornetWorkspaceDescriptor_t &workDesc,
-                            void *scratchPtr, std::size_t worksize) {
+                            void *scratchPtr, std::size_t &worksize) {
         PL_CUTENSORNET_IS_SUCCESS(cutensornetWorkspaceSetMemory(
             /* const cutensornetHandle_t */ getTNCudaHandle(),
             /* cutensornetWorkspaceDescriptor_t */ workDesc,
@@ -207,7 +343,7 @@ class TNCudaBase : public TensornetBase<Precision, Derived> {
 
         std::size_t worksize = getWorkSpaceMemorySize(workDesc);
 
-        PL_ABORT_IF(std::size_t(worksize) > scratchSize,
+        PL_ABORT_IF(worksize > scratchSize,
                     "Insufficient workspace size on Device!");
 
         const std::size_t d_scratch_length = worksize / sizeof(std::size_t);